requirements.txt
grandalf
pypdf
langchain>=0.1.16
langchain-core>=0.3.72
langchain-openai>=0.0.8
langgraph>=0.0.36
langchain-groq
langchain-anthropic
langchain-google-genai
langchain-community
python-dotenv
linkedin-api
streamlit
nest-asyncio
streamlit_chat
streamlit_pills
langsmith
python-docx
pydantic>=2.0.0
pandas